The lvdrawduf.{h,cpp} files are splited according to the principle of one class to a pair of files {h,cpp}. This files placed into crengine/src/lvdrawbuf.
