+++ android/jni/cr3java.cpp
@@ -1,5 +1,6 @@
 #include "cr3java.h"
-#include "../../crengine/include/crlog.h"
+#include "crlog.h"
+#include "lvcolordrawbuf.h"
 
 #include <dlfcn.h>
 #include <android/api-level.h>
