+++ crengine/src/crskin.cpp
@@ -15,6 +15,7 @@
 #include "../include/crskin.h"
 #include "../include/lvstsheet.h"
 #include "../include/crtrace.h"
+#include "../include/lvdrawstatesaver.h"
 
 // uncomment to trace skin XML access errors / not found elements
 //#define TRACE_SKIN_ERRORS
